OH MY GOD I GOT IT TO WORK
OH MY GOD OH MY GOD UNIVERSAL PLAYBACK ACTUALLY WORKS IN SOME SORT
