+++ client/src/components/AudioPlayer.css
@@ -1,0 +1,15 @@
+.audio-player {
+    position: fixed;       /* Fix it to the bottom of the screen */
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    z-index: 9999;         /* Ensure it's on top of other elements */
+    background-color: #fff;
+    padding: 0.5rem;
+    box-shadow: 0 -2px 5px rgba(0, 0, 0, 0.2); /* Optional shadow above the player */
+  }
+  
+  .audio-player audio {
+    width: 100%;           /* Stretch the audio controls across the full width */
+  }
+
