+++ client/src/contexts/AudioContext.js
@@ -1,0 +1,15 @@
+// AudioContext.js
+import React, { createContext, useState, useContext } from 'react';
+
+const AudioContext = createContext();
+
+export const AudioProvider = ({ children }) => {
+  const [currentSong, setCurrentSong] = useState(null);
+  return (
+    <AudioContext.Provider value={{ currentSong, setCurrentSong }}>
+      {children}
+    </AudioContext.Provider>
+  );
+};
+
+export const useAudio = () => useContext(AudioContext);
